feat(xlsx): support primitive arrays and index markers
- Enable [[array]] markers for primitive value interpolation
- Add special index markers: $index, $index1, and $number
- Update documentation in README.md and AGENTS.md
- Add functional tests and update Spark journal
