+++ .axioma/spark.md
@@ -1,1 +1,5 @@
 # Spark's Journal
+
+## 2025-05-15 - [XLSX Template Enhancements]
+**Learning:** Users often need to render simple lists or include row counters in their documents. Adding support for primitive arrays and special index markers significantly improves the library's versatility with minimal code changes.
+**Pattern:** Using special property paths (like `$index`, `$number`) is an effective way to expose metadata about the current iteration without cluttering the input data object.
